fetch/api/basic/block-mime-as-script.html

<!doctype html>
<meta charset="utf-8">
<title>Block mime type as script</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div></div>
<script>
  const noop = () => {};
  const expect_error = (e) => {
    assert_equals(e.constructor, Event);
  };

  ["non-empty", "empty"].forEach(function(content) {
    ["text/csv",
     "audio/aiff",
     "audio/midi",
     "audio/whatever",
     "video/avi",
     "video/fli",
     "video/whatever",
     "image/jpeg",
     "image/gif",
     "image/whatever"].forEach(function(test_case) {
      const url = "../resources/script-with-header.py?content=" + content +
                  "&mime=" + test_case;

      async_test(function(t) {
        var script = document.createElement("script");
        script.onerror = t.step_func_done(expect_error);
        script.onload = t.unreached_func("Unexpected load event");
        script.src = url;
        document.body.appendChild(script);
      }, "Should fail loading " + content + " script with " + test_case +
         " MIME type (script)");

      async_test(function(t) {
        var script = document.createElement("script");
        script.setAttribute("type", "module");
        script.onerror = t.step_func_done(expect_error);
        script.onload = t.unreached_func("Unexpected load event");
        script.src = url;
        document.body.appendChild(script);
      }, "Should fail loading " + content + " script with " + test_case +
         " MIME type (module script)");

      async_test(function(t) {
        const worker = new Worker(url);
        worker.onerror = t.step_func_done(expect_error);
        worker.onload = t.unreached_func("Unexpected load event");
      }, "Should fail loading " + content + " script with " + test_case +
         " MIME type (dedicated worker)");

      async_test(function(t) {
        const worker = new Worker(url, {type: "module"});
        worker.onerror = t.step_func_done(expect_error);
        worker.onload = t.unreached_func("Unexpected load event");
      }, "Should fail loading " + content + " script with " + test_case +
         " MIME type (module dedicated worker)");
    });
  });

  ["html", "plain"].forEach(function(test_case) {
    async_test(function(t) {
      var script = document.createElement("script");
      script.onerror = t.unreached_func("Unexpected error event");
      script.onload = t.step_func_done(noop);
      script.src = "../resources/script-with-header.py?mime=text/" + test_case;
      document.body.appendChild(script);
    }, "Should load script with text/" + test_case + " MIME type (script)");
  });

</script>
